Added UnpackReader which puts the eventno and trigger into the root file correctly

+++ c4data/germaniumData/GermaniumFebexData.h
@@ -16,7 +16,7 @@ class GermaniumFebexData : public TObject
                              uint8_t pileup,
                              uint8_t overflow,
                              uint8_t channel_id,
-                             uint64_t channel_trigger_time,
+                             double channel_trigger_time,
                              int32_t channel_energy,
                              double channel_energy_cal,
                              uint32_t crystal_id,
@@ -38,7 +38,7 @@ class GermaniumFebexData : public TObject
         inline const uint8_t Get_pileup() const { return fpileup; }
         inline const uint8_t Get_overflow() const { return foverflow; }
         inline const uint8_t Get_channel_id() const { return fchannel_id; }
-        inline const uint64_t Get_channel_trigger_time() const { return fchannel_trigger_time; }
+        inline const double Get_channel_trigger_time() const { return fchannel_trigger_time; }
         inline const int32_t Get_channel_energy() const { return fchannel_energy; }
         inline const double Get_channel_energy_cal() const { return fchannel_energy_cal; }
         inline const uint32_t Get_crystal_id() const { return fcrystal_id; }
@@ -56,7 +56,7 @@ class GermaniumFebexData : public TObject
         void Set_pileup(uint8_t v) { fpileup = v; }
         void Set_overflow(uint8_t v) { foverflow = v; }
         void Set_channel_id(uint8_t v) { fchannel_id = v; }
-        void Set_channel_trigger_time(uint64_t v) { fchannel_trigger_time = v; }
+        void Set_channel_trigger_time(double v) { fchannel_trigger_time = v; }
         void Set_channel_energy(int32_t v) { fchannel_energy = v; }
         void Set_channel_energy_cal(double v) { fchannel_energy_cal = v; }
         void Set_crystal_id(uint32_t v) { fcrystal_id = v; }
@@ -74,9 +74,9 @@ class GermaniumFebexData : public TObject
         uint8_t fpileup;
         uint8_t foverflow;
         uint8_t fchannel_id;
-        uint64_t fchannel_trigger_time;
+        double fchannel_trigger_time;
         int32_t fchannel_energy;
-        int32_t fchannel_energy_cal;
+        double fchannel_energy_cal;
         uint32_t fcrystal_id;
         uint32_t fdetector_id;
